:bug: handle subprocess exceptions and remove notifications
AppleScript notifications have been removed, as it was difficult
to keep track of song changes to know when to fire off a notification.
This commit also fixes a crash that occurred when the AppleScript
commands failed/exited with non-zero exit code.
Lastly, a "No track playing" message has been added.